Move metadata to own file
