+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/jobs/catalog.py
@@ -2,7 +2,7 @@ from dagster import define_asset_job
 
 
 generate_catalog = define_asset_job(
-    name="generate_catalog", selection=["metadata_directory_report"]
+    name="generate_catalog", selection=["metadata_directory_report", "metadata_definitions"]
 )
 generate_catalog_markdown = define_asset_job(
     name="generate_catalog_markdown", selection=["connector_catalog_location_html", "connector_catalog_location_markdown"]
